TASK-51 [Refactor] : 파일명 변경, aria-label 추가 등 코드 리뷰 반영
