+++ src/components/edit/InvitationList.tsx
@@ -3,7 +3,7 @@
 import Image from 'next/image';
 
 import Pagination from '@/src/components/edit/Pagination';
-import usePagination from '@/src/lib/hooks/usePaginavigation';
+import usePagination from '@/src/lib/hooks/usePagination';
 
 // 테스트용 더미 데이터
 interface Item {
