+++ src/components/edit/Pagination.tsx
@@ -39,7 +39,11 @@ export default function Pagination({
         {totalPages} 페이지 중 {currentPage}
       </div>
       <div>
-        <button type="button" onClick={handlePrePage}>
+        <button
+          type="button"
+          onClick={handlePrePage}
+          aria-label="이전 페이지로 이동"
+        >
           <div
             className="flex h-10 w-10 items-center justify-center border border-gray-200"
             style={{
@@ -57,7 +61,11 @@ export default function Pagination({
             />
           </div>
         </button>
-        <button type="button" onClick={handleNextPage}>
+        <button
+          type="button"
+          onClick={handleNextPage}
+          aria-label="다음 페이지로 이동"
+        >
           <div
             className="flex h-10 w-10 items-center justify-center border border-gray-200"
             style={{
